+++ Samples/Sample-Copy-SSHFolder.ps1
@@ -1,5 +1,10 @@
 
+$modulePath = "D:\Projects\PSNovusTools\PSNovusTools"
+Remove-Module -Name $modulePath -Force
+Import-Module -Name $modulePath -Force
 
+<#
+Get-Module -Name PSNovusTools
 
 $module = Get-Module -Name PSNovusTools
 
@@ -7,5 +12,5 @@ if (-not $module) {
     Write-Host "PSNovusTools module is not installed. "
    exit
 } 
-
-Copy-SSHFolder -remoteServer 10.0.0.21 -remoteUser Administrator -remotePassword (Convert-StringToSecureString("w0rk1ngh@rd")) -localFolderPath "D:\Projects\PSNovusTools\Samples" -remoteFolderPath "c:\temp"
+#>
+Copy-SSHFolder -remoteServer 10.0.0.21 -remoteUser Administrator -remotePassword (Convert-StringToSecureString("w0rk1ngh@rd")) -localFolderPath "D:\Projects\PSNovusTools\Samples" -remoteFolderPath "c:\temp" -Verbose