+++ Changelog.md
@@ -3,17 +3,24 @@
 <!-- TOC -->
 
 - [Changelog](#changelog)
+    - [Release v0.1.3](#release-v013)
     - [Release v0.1.2](#release-v012)
     - [Release v0.1.1](#release-v011)
     - [Release v0.1.0](#release-v010)
 
 <!-- /TOC -->
 
+## Release v0.1.3
+
+* New private function Get-SSHRemoteOS 
+* Added Verbose to function Copy-SSHFolder
+* Updated function Test-Windows2022Or11OrHigher with correct Verbose switch
+
 ## Release v0.1.2
 
 * New function Copy-SSHFolder
 * Update Copy-SSHFile with new Close-SSHSession
-* New private Close-SSHSession function 
+* New private function Close-SSHSession 
 
 
 ## Release v0.1.1
@@ -28,3 +35,12 @@
 
 * First published to powershellgallery.com
 
+<!-- TOC -->
+
+- [Changelog](#changelog)
+    - [Release v0.1.3](#release-v013)
+    - [Release v0.1.2](#release-v012)
+    - [Release v0.1.1](#release-v011)
+    - [Release v0.1.0](#release-v010)
+
+<!-- /TOC -->